Add separate module for createAlpsFile function
Extracted the createAlpsFile function from src/extension.ts into its own module for better code organization and maintainability. This change helps isolate the functionality, making the codebase easier to understand and extend.
